clean up theme implementation
